+++ AElf.Kernel/Consensus/DPoS.cs
@@ -1,0 +1,16 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace AElf.Kernel
+{
+    public class DPoS : IMiner
+    {
+        public byte[] Mine(IBlockHeader blockheader)
+        {
+            throw new NotImplementedException();
+        }
+
+
+    }
+}
