+++ AElf.Kernel/Consensus/IDelegate.cs
@@ -1,0 +1,7 @@
+﻿namespace AElf.Kernel
+{
+    public interface IDelegate
+    {
+        void AddDelegate();
+    }
+}
