--- AElf.Kernel/IDelegates.cs
@@ -1,7 +1,0 @@
-﻿namespace AElf.Kernel
-{
-    public interface IDelegates
-    {
-
-    }
-}
