--- AElf.Kernel/IMerkleTree.cs
@@ -1,8 +1,0 @@
-﻿namespace AElf.Kernel
-{
-	public interface IMerkleTree<T>
-	{
-		IHash<IMerkleTree<T>> ComputeRootHash();
-		void AddNode(IHash<T> hash);
-	}
-}
