+++ AElf.Kernel/MerkleTree/IMerkleTree.cs
@@ -1,0 +1,8 @@
+﻿namespace AElf.Kernel
+{
+    public interface IMerkleTree<T>
+    {
+        IHash<IMerkleTree<T>> ComputeRootHash();
+        void AddNode(IHash<T> hash);
+    }
+}
