+++ AElf.Kernel/MerkleTree/MerkleNode.cs
